family size: 0.5-step servings in planner, hint in recipe edit form
- week_plan.servings changed from INTEGER to REAL (migration 6f7a8b9c0d1e)
- All servings API endpoints use float instead of int
- Wochenplan loads familySize from settings; new recipes default to familySize
- Servings stepper uses ±0.5 steps, min 0.5
- Recipe edit form label shows "(Mengen für 1 Person)"

+++ app/models/week_plan.py
@@ -11,7 +11,7 @@ class WeekPlan(db.Model):
                      db.ForeignKey('recipes.id', ondelete='SET NULL'),
                      nullable=True)
 
-    servings   = db.Column(db.Integer, nullable=False, default=1)
+    servings   = db.Column(db.Float, nullable=False, default=1.0)
     is_bought  = db.Column(db.Boolean, nullable=False, default=False)
 
     recipe = db.relationship('Recipe',
